Make BuildAction conform ModeConvertible
